+++ classes/Admin/AdminController.php
@@ -36,6 +36,7 @@ use Nyholm\Psr7\ServerRequest;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use RocketTheme\Toolbox\Event\Event;
+use RocketTheme\Toolbox\ResourceLocator\UniformResourceLocator;
 use RocketTheme\Toolbox\Session\Message;
 
 /**
@@ -338,17 +339,39 @@ class AdminController
         }
 
         $data = $this->data;
+        $route = trim($data['route'] ?? '', '/');
 
-        if ($data['route'] === '' || $data['route'] === '/') {
-            $path = $this->grav['locator']->findResource('page://');
-        } else {
-            $path = $this->grav['page']->find($data['route'])->path();
+        // TODO: Folder name needs to be validated!
+        $folder = mb_strtolower($data['folder'] ?? '');
+        if ($folder === '') {
+            throw new \RuntimeException('Creating folder failed, bad folder name', 400);
         }
 
-        $orderOfNewFolder = ''; //static::getNextOrderInFolder($path) . '.';
-        $new_path         = $path . '/' . $orderOfNewFolder . $data['folder'];
+        /** @var PageObject $parent */
+        $parent = $route ? $directory->getObject($route) : $directory->getCollection()->getRoot();
+        if (!$parent) {
+            throw new \RuntimeException('Creating folder failed, bad parent route', 400);
+        }
 
-        Folder::create($new_path);
+        $path = $parent->getFlexDirectory()->getStorageFolder($parent->getStorageKey());
+        if (!$path) {
+            throw new \RuntimeException('Creating folder failed, bad parent storage path', 400);
+        }
+
+        // Ordering
+        $orders = $parent->children()->visible()->getProperty('order');
+        $maxOrder = 0;
+        foreach ($orders as $order) {
+            $maxOrder = max($maxOrder, (int)$order);
+        }
+
+        $orderOfNewFolder = $maxOrder ? sprintf('%02d.', $maxOrder+1) : '';
+        $new_path         = $path . '/' . $orderOfNewFolder . $folder;
+
+        /** @var UniformResourceLocator $locator */
+        $locator = $this->grav['locator'];
+
+        Folder::create($locator->findResource($new_path, true, true));
         Cache::clearCache('invalidate');
 
         $this->grav->fireEvent('onAdminAfterSaveAs', new Event(['path' => $new_path]));